Flatten Support::Requests::Anonymous module for models
This change doesn't make any functional modifications. What's changed:

- the `Support::Requests::Anonymous` module has been removed because
it made the directory structure hard to work with and caused lots of
typing on the rails console
- require statements featuring models located in `app/models` have been
removed (Rails autoloads those classes automatically)
- two new model files have been created (`TopUrls`, `FeedbackCounts`)
for classes that were housed within the file belonging to another class.
It's best practice for each ruby file to only contain one class.
