--- app/models/support/requests/anonymous.rb
@@ -1,5 +1,0 @@
-module Support::Requests::Anonymous; end
-
-require 'support/requests/anonymous/long_form_contact'
-require 'support/requests/anonymous/problem_report'
-require 'support/requests/anonymous/service_feedback'
