+++ app/workers/problem_report_worker.rb
@@ -1,10 +1,8 @@
-require 'support/requests/anonymous/problem_report'
-
 class ProblemReportWorker
   include Sidekiq::Worker
 
   def perform(problem_report_params)
-    problem_report = Support::Requests::Anonymous::ProblemReport.create!(problem_report_params)
+    problem_report = ProblemReport.create!(problem_report_params)
     ZendeskTicketWorker.perform_async(problem_report.id)
     ContentItemEnrichmentWorker.perform_async(problem_report.id)
   end
