+++ app/workers/service_feedback_worker.rb
@@ -1,9 +1,7 @@
-require 'support/requests/anonymous/service_feedback'
-
 class ServiceFeedbackWorker
   include Sidekiq::Worker
 
   def perform(service_feedback_params)
-    Support::Requests::Anonymous::ServiceFeedback.new(service_feedback_params).save!
+    ServiceFeedback.new(service_feedback_params).save!
   end
 end
