+++ spec/factories/anonymous_feedback.rb
@@ -1,21 +1,17 @@
-require 'support/requests/anonymous/anonymous_contact'
-require 'support/requests/anonymous/problem_report'
-require 'support/requests/anonymous/service_feedback'
-
 FactoryGirl.define do
-  factory :anonymous_contact, class: Support::Requests::Anonymous::AnonymousContact do
+  factory :anonymous_contact, class: AnonymousContact do
     javascript_enabled true
     path "/vat-rates"
 
-    factory :problem_report, class: Support::Requests::Anonymous::ProblemReport
+    factory :problem_report, class: ProblemReport
 
-    factory :service_feedback, class: Support::Requests::Anonymous::ServiceFeedback do
+    factory :service_feedback, class: ServiceFeedback do
       path { "/done/#{slug}" }
       slug "apply-carers-allowance"
       service_satisfaction_rating 5
     end
 
-    factory :long_form_contact, class: Support::Requests::Anonymous::LongFormContact
+    factory :long_form_contact, class: LongFormContact
   end
 
   factory :organisation do
